Add PlayerRepository, Refactor Game
